+++ pycognaize/document/field/numeric_field.py
@@ -31,7 +31,8 @@ class NumericField(Field):
                  confidence: Optional[float] = -1.0,
                  group_name: str = None,
                  scale: int = None,
-                 mapping: Optional[List[Dict[str, str]]] = None
+                 mapping: Optional[List[Dict[str, str]]] = None,
+                 is_calculated: bool = None
                  ):
         super().__init__(name=name, tags=tags, value=value,
                          group_key=group_key, confidence=confidence,
@@ -49,6 +50,7 @@ class NumericField(Field):
             self._value = sum([self.convert_to_numeric(i.raw_value)
                                for i in self.tags])
             self._tag_value = self._value
+        self._is_calculated = is_calculated
 
     @property
     def name(self):
@@ -74,6 +76,10 @@ class NumericField(Field):
     def raw_field_value(self):
         return self._raw_field_value
 
+    @property
+    def is_calculated(self):
+        return self._is_calculated
+
     @classmethod
     def construct_from_raw(cls, raw: dict, pages: Dict[int, Page],
                            html: Optional[HTML] = None,
@@ -95,8 +101,11 @@ class NumericField(Field):
         field_value = raw[IqFieldKeyEnum.value.value]
         field_value = (tags[0].raw_value if (html.path and tags)
                        else field_value)
+        is_calculated = raw.get(IqFieldKeyEnum.field_type.value, '') == "extraction and computation"
+
         return cls(name=raw[IqDocumentKeysEnum.name.value],
                    value=field_value,
+                   is_calculated=is_calculated,
                    calculated_value=calculated_value,
                    tags=tags,
                    field_id=str(raw[ID]),